+++ src/hooks/useMapInteraction.ts
@@ -298,6 +298,7 @@ export function useMapInteraction(
           { layers: ["street-trees-national-circle"], type: "streetTreesNational" },
           { layers: ["tree-pits-taipei-fill", "tree-pits-taipei-line"], type: "treePitsTaipei" },
           { layers: ["buildings-gba-fill", "buildings-gba-extrusion"], type: "buildingsGba" },
+          { layers: ["urban-form-grid-fill"], type: "urbanFormGrid" },
           { layers: ["sports-venues-school", "sports-venues-public-other", "sports-venues-private", "sports-venues-park", "sports-venues-center"], type: "sportsVenue" },
           { layers: ["sat-current-point"], type: "satellite" },
           { layers: ["medical-hospitals-circle"], type: "medicalPOI" },